Changes capitalized notification types, moved the types to the type folder and fixed sonar issues with same names.
