+++ src/types/SystemMessageNotification.ts
@@ -1,0 +1,1 @@
+export type SystemMessageNotification = 'Info' | 'Warning' | 'Notice';
